move slider code to separate file
