+++ src/pages/Query.jsx
@@ -1,13 +1,16 @@
 import React from "react";
 import { Helmet } from 'react-helmet';
 import { Select } from "react-dropdown-select";
-import * as d3 from 'd3';
-import { createD3RangeSlider } from '../SDK/d3RangeSlider.js';
 import QueryResult from '../components/QueryResult';
 import QueryIntro from "../components/QueryIntro";
 import Paginator from '../components/Paginator';
 import DataReference from '../components/DataReference';
 import { useLocation } from 'react-router-dom';
+import {
+    drawSliders,
+    sliderIdentity,
+    sliderCoverage
+} from '../helpers/QuerySliders';
 import {
     getPlaceholder,
     getPageLinks,
@@ -210,51 +213,3 @@ const Query = (props) => {
 }
 
 export default Query;
-
-var sliderIdentity;
-var sliderCoverage;
-
-var identityLims = [75, 100];
-var coverageLims = [25, 100];
-var coverageColorLims = ["#3d5088", "#fce540"];
-
-var sliderIdentityLabelL;
-var sliderIdentityLabelR;
-var sliderCoverageLabelL;
-var sliderCoverageLabelR;
-
-const drawSliders = () => {
-    var sliderIdentityDiv = d3.select("#sliderIdentity");
-    var sliderCoverageDiv = d3.select("#sliderCoverage");
-
-    sliderIdentity = createD3RangeSlider(d3, identityLims[0], identityLims[1], sliderIdentityDiv);
-    sliderIdentity.onChange((range) => updateIdentityLims(range.begin, range.end));
-
-    var coverageColorGradient = `background-image: linear-gradient(to right, ${coverageColorLims[0]} , ${coverageColorLims[1]});`
-    var newCoverageSliderDivStyle = sliderCoverageDiv.attr("style") + coverageColorGradient;
-    sliderCoverage = createD3RangeSlider(d3, coverageLims[0], coverageLims[1], sliderCoverageDiv);
-    sliderCoverage.onChange((range) => updateCoverageLims(range.begin, range.end));
-    sliderCoverageDiv.attr("style", newCoverageSliderDivStyle)
-    sliderCoverageDiv.select(".slider-container")
-        .attr("style", coverageColorGradient)
-    sliderCoverageDiv.select(".slider")
-        .attr("style", "background: rgba(0,0,0, 0.2)")
-
-    sliderIdentityLabelL = sliderIdentityDiv.select(".WW").append("span")
-        .attr("style", "float: left; transform: translate(0px,20px)");
-    sliderIdentityLabelR = sliderIdentityDiv.select(".EE").append("text")
-        .attr("style", "float: left; transform: translate(-5px,20px)");
-    sliderCoverageLabelL = sliderCoverageDiv.select(".WW").append("span")
-        .attr("style", "float: left; transform: translate(0px,20px)");
-    sliderCoverageLabelR = sliderCoverageDiv.select(".EE").append("text")
-        .attr("style", "float: left; transform: translate(-5px,20px)");
-};
-
-const updateIdentityLims = (begin, end) => {
-    sliderIdentityLabelL.text(begin);
-    sliderIdentityLabelR.text(end);
-};
-const updateCoverageLims = (begin, end) => {
-    sliderCoverageLabelL.text(begin);
-    sliderCoverageLabelR.text(end);
-};
